\name{pcse-package}
\alias{pcse-package}
\docType{package}
\title{
\packageTitle{pcse}
}
\description{
\packageDescription{pcse}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pcse}
\packageIndices{pcse}

}
\author{
\packageAuthor{pcse}

Maintainer: \packageMaintainer{pcse}
}
\references{
Alvarez, R. Michael, Geoffrey Garrett and Peter Lange. (1991) Government Partisanship, Labor Organization, and Macroeconomic Performance. \emph{American Political Science Review} 85:539-56.

Beck, Nathaniel, Jonathan N. Katz, R. Michael Alvarez, Geoffrey Garrett and Peter Lange. (1993) Government Partisanship, Labor Organization, and Macroeconomic Performance: A Corrigendum. \emph{American Political Science Review} 87:945-948.

Beck, Nathaniel and Jonathan N. Katz. (1995) What to do (and not to do) with Time-Series Cross-Section Data. \emph{American Political Science Review} 89:634-647.

Bailey, Delia and Jonathan N. Katz. (2011) Implementing Panel-Corrected Standard Errors in R: The pcse Package. \emph{Journal of Statistical Software, Code Snippets} 42(1):1--11. \url{http://www.jstatsoft.org/v42/c01/}
}

\keyword{ package }
\examples{
## see demo file.
}
