\name{summary.pcse}
\alias{summary.pcse}
\title{Summary Method for Package pcse}
\description{
The package pcse contains a function to estimate panel-corrected standard errors. Data may contain balanced or unbalanced panels. This function summarizes the estimated results.
}

\usage{\method{summary}{pcse}(object, ...)}
\arguments{
\item{object}{An object of class "pcse."}
\item{...}{Arguments passed to other functions.}
}


\examples{
## see demo file.
}
\references{
Bailey, Delia and Jonathan N. Katz. (2011) Implementing Panel-Corrected Standard Errors in R: The pcse Package. \emph{Journal of Statistical Software, Code Snippets} 42(1):1--11. \url{http://www.jstatsoft.org/v42/c01/}
}

\author{
Delia Bailey <delia.bailey@gmail.com> and Jonathan N. Katz <jkatz@caltech.edu>
Maintainer: Delia Bailey <delia.bailey@gmail.com>
}

\keyword{file}


