\name{summary.pcse}
\alias{summary.pcse}
\title{Summary Method for Package pcse}
\description{
The package pcse contains a function to estimate panel-corrected standard errors. Data may contain balanced or unbalanced panels. This function summarizes the estimated results.
}

\usage{\method{summary}{pcse}(object, ...)}
\arguments{
\item{object}{An object of class "pcse."}
\item{...}{Arguments passed to other functions.}
}


\examples{
## see demo file.
}
\references{
}

\author{
Delia Bailey <delia@caltech.edu> and Jonathan N. Katz <jkatz@caltech.edu>
Maintainer: Delia Bailey <delia@caltech.edu>
}

\keyword{file}


